moving stuff around
dfa2matcher and regex2nfa are not passes
