+++ gtest/automaton.cpp
@@ -10,10 +10,8 @@
 
 #include <thorin/fe/parser.h>
 
-#include <thorin/plug/regex/pass/regex2nfa.h>
-#include <thorin/plug/regex/regex.h>
-
-#include "thorin/plug/regex/pass/dfa2matcher.h"
+#include <thorin/plug/regex/dfa2matcher.h>
+#include <thorin/plug/regex/regex2nfa.h>
 
 using namespace automaton;
 using namespace thorin;
