+++ plugins/regex/CMakeLists.txt
@@ -1,15 +1,15 @@
 add_thorin_plugin(regex
     SOURCES
         src/thorin/plug/regex/normalizers.cpp
+        src/thorin/plug/regex/dfa2matcher.cpp
         src/thorin/plug/regex/regex.cpp
-        src/thorin/plug/regex/pass/dfa2matcher.cpp
+        src/thorin/plug/regex/regex2nfa.cpp
         src/thorin/plug/regex/pass/lower_regex.cpp
-        src/thorin/plug/regex/pass/regex2nfa.cpp
     HEADERS
+        include/thorin/plug/regex/dfa2matcher.h
         include/thorin/plug/regex/regex.h
-        include/thorin/plug/regex/pass/dfa2matcher.h
+        include/thorin/plug/regex/regex2nfa.h
         include/thorin/plug/regex/pass/lower_regex.h
-        include/thorin/plug/regex/pass/regex2nfa.h
     INTERFACE
         automaton
         thorin_interface_core
